pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
authors = [{ "name" = "Tom Carrick", "email" = "tom@carrick.eu" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = ["django", "django-q2"]
description = "An asynchronous Django email backend for Django Q2"
license = { "file" = "LICENSE" }
name = "django-q2-email-backend"
readme = "README.rst"
requires-python = ">=3.10"
version = "1.0.0"

[dependency-groups]
test = ["coverage[toml]", "django-stubs", "ruff", "ty"]

[project.urls]
Homepage = "https://github.com/knyghty/django-q2-email-backend"
Repository = "https://github.com/knyghty/django-q2-email-backend"

[tool.coverage.run]
branch = true
source = ["."]

[tool.hatch.build.targets.sdist]
exclude = ["/tests", "requirements.txt", "runtests.py"]

[tool.hatch.build.targets.wheel]
packages = ["django_q2_email_backend"]

[tool.ruff.lint]
ignore = ["ISC001"]
select = [
    "ANN",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "C90",
    "DJ",
    "E",
    "ERA",
    "EXE",
    "F",
    "FA",
    "FLY",
    "G",
    "I",
    "INP",
    "INT",
    "ISC",
    "N",
    "PERF",
    "PGH",
    "PIE",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLOT",
    "T10",
    "T20",
    "TCH",
    "TRY",
    "UP",
    "W",
    "YTT",
]

[tool.ruff.lint.isort]
force-single-line = true
